Refactor displays into modules and remove DearPyGui
