+++ display/__init__.py
@@ -1,0 +1,9 @@
+from .interface import IDisplay
+from .turtle_display import TurtleDisplay
+from .tkinter_display import TkinterDisplay
+
+__all__ = [
+    "IDisplay",
+    "TurtleDisplay",
+    "TkinterDisplay",
+]
