+++ display/interface.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+
+from graph import Graphable
+
+class IDisplay(ABC):
+    @abstractmethod
+    def drawGraph(self, graph: Graphable) -> None:
+        ...
