--- requirements.txt
@@ -1,1 +1,0 @@
-dearpygui==1.10.0
